fix(postgres): use long arithmetic for pagination offset to avoid int overflow

+++ Adaptors/PostgresSQL/src/PartitionTable.cs
@@ -211,7 +211,7 @@ INSERT INTO partitions (
     dataCmd.Parameters.AddWithValue("limit",
                                     pageSize);
     dataCmd.Parameters.AddWithValue("offset",
-                                    page * pageSize);
+                                    (long)page * pageSize);
 
     await using var reader = await dataCmd.ExecuteReaderAsync(cancellationToken)
                                           .ConfigureAwait(false);
